Changed the Layout of the Website
Added a responsive top navigation menu, added a Hero section, Re-located the image and title of the "About Me" section and added a functional button where it will direct to another link. Changed the "My Works" project layouts, but kept the hovering effect. I just added a picture at the "Contact" section to maximize the space.